move hybrid code into hybrid dir/namespace
