+++ src/ngraph/pass/assign_placement.hpp
@@ -32,13 +32,9 @@ namespace ngraph
         public:
             // TODO: make policy a class
             AssignPlacement(std::function<Placement(std::shared_ptr<Node>)> placement_policy);
-            AssignPlacement(
-                std::vector<std::shared_ptr<ngraph::runtime::Backend>> placement_backends);
 
         private:
             bool run_on_node(std::shared_ptr<Node> node) override;
-
-            std::vector<std::shared_ptr<ngraph::runtime::Backend>> m_placement_backends;
 
             std::function<Placement(std::shared_ptr<Node>)> m_placement_policy;
         };
