+++ src/ngraph/runtime/hybrid/hybrid_backend.cpp
@@ -16,9 +16,9 @@
 
 #include "ngraph/runtime/hybrid/hybrid_backend.hpp"
 #include "ngraph/graph_util.hpp"
-#include "ngraph/pass/assign_placement.hpp"
 #include "ngraph/pass/manager.hpp"
 #include "ngraph/runtime/hybrid/hybrid_util.hpp"
+#include "ngraph/runtime/hybrid/pass/assign_placement.hpp"
 #include "ngraph/runtime/tensor.hpp"
 
 using namespace ngraph;
@@ -81,8 +81,8 @@ runtime::Handle runtime::hybrid::HybridBackend::compile(shared_ptr<Function> fun
         instance.m_function = clone_function(*func);
 
         // Run placement pass
-        pass::Manager pass_manager;
-        pass_manager.register_pass<pass::AssignPlacement>(backend_list);
+        ngraph::pass::Manager pass_manager;
+        pass_manager.register_pass<runtime::hybrid::pass::AssignPlacement>(backend_list);
         pass_manager.run_passes(instance.m_function);
 
         // Split function to sub_functions
